RecyclerView뷰에서 ViewModel자체를 바인딩 하지 않도록 변경 & BindingAdapter 위치 binding 패키지로 변경
